+++ include/e3e_detection.h
